Install DALI, load data with it, change to pytorch 0.4
For compability with DALI, pytorch 0.4 is required,
docker container was updated to the first one that
supports it, and flownet2-pytorch was updated with
suggestions from warings for compability.

Custom DALI loader was added. It works simillary
to sequence loader from pytorch.

Input from that loader is further processed by numpy,
for now DALI only loads sequences without doing augmentations
(that is required cropping and channel transpose).